requirements.txt
streamlit>=1.32.0
sqlalchemy>=2.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
asyncpg==0.29.0
psycopg[binary]
langchain>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.40
langchain-community>=0.0.20
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
numpy>=1.24.0
passlib[bcrypt]>=1.7.4
huggingface-hub>=0.16.0
transformers>=4.21.0